refactor: 포트 어댑터로 도메인 간 의존 분리
