+++ src/main/java/org/sopt/domain/post/application/port/UserPort.java
@@ -1,0 +1,8 @@
+package org.sopt.domain.post.application.port;
+
+import org.sopt.domain.user.domain.model.User;
+
+public interface UserPort {
+
+    User getUser(Long userId);
+}
