+++ src/main/java/org/sopt/domain/user/application/port/PostPort.java
@@ -1,0 +1,6 @@
+package org.sopt.domain.user.application.port;
+
+public interface PostPort {
+
+    boolean existsByAuthorUserId(Long authorUserId);
+}
